FEATURE: re-implement node:migrationcreate command:
